requirements.txt
flask
openai
python-dotenv
requests
gunicorn
PyPDF2